Refactoring io.delta package to io.delta.tables
Staging this PR for now.
Changing the namespace `io.delta` to `io.delta.tables`

No new tests added, re ran old tests.

Author: Rahul Mahadev <rahul.mahadev@databricks.com>

GitOrigin-RevId: ab6a5968a29f698426bea27db0f33f431da637ef
