+++ src/main/scala/io/delta/tables/execution/DeltaTableOperations.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.delta.execution
+package io.delta.tables.execution
 
 import scala.collection.Map
 
@@ -22,7 +22,7 @@ import org.apache.spark.sql.delta.PreprocessTableUpdate
 import org.apache.spark.sql.delta.{DeltaErrors, DeltaFullTable, DeltaHistoryManager, DeltaLog}
 import org.apache.spark.sql.delta.commands.{DeleteCommand, VacuumCommand}
 import org.apache.spark.sql.delta.util.AnalysisHelper
-import io.delta.DeltaTable
+import io.delta.tables.DeltaTable
 
 import org.apache.spark.sql.{functions, Column, DataFrame, SparkSession}
 import org.apache.spark.sql.catalyst.analysis.EliminateSubqueryAliases
