+++ src/test/scala/io/delta/tables/DeltaTableSuite.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.delta
+package io.delta.tables
 
 import java.util.Locale
