+++ src/test/scala/org/apache/spark/sql/delta/DeleteScalaSuite.scala
@@ -16,7 +16,7 @@
 
 package org.apache.spark.sql.delta
 
-import io.delta.DeltaTable
+import io.delta.tables.DeltaTable
 
 import org.apache.spark.sql.{functions, Row}
 
@@ -26,21 +26,21 @@ class DeleteScalaSuite extends DeleteSuiteBase {
 
   test("delete usage test - without condition") {
     append(Seq((1, 10), (2, 20), (3, 30), (4, 40)).toDF("key", "value"))
-    val table = io.delta.DeltaTable.forPath(tempPath)
+    val table = io.delta.tables.DeltaTable.forPath(tempPath)
     table.delete()
     checkAnswer(readDeltaTable(tempPath), Nil)
   }
 
   test("delete usage test - with condition") {
     append(Seq((1, 10), (2, 20), (3, 30), (4, 40)).toDF("key", "value"))
-    val table = io.delta.DeltaTable.forPath(tempPath)
+    val table = io.delta.tables.DeltaTable.forPath(tempPath)
     table.delete("key = 1 or key = 2")
     checkAnswer(readDeltaTable(tempPath), Row(3, 30) :: Row(4, 40) :: Nil)
   }
 
   test("delete usage test - with Column condition") {
     append(Seq((1, 10), (2, 20), (3, 30), (4, 40)).toDF("key", "value"))
-    val table = io.delta.DeltaTable.forPath(tempPath)
+    val table = io.delta.tables.DeltaTable.forPath(tempPath)
     table.delete(functions.expr("key = 1 or key = 2"))
     checkAnswer(readDeltaTable(tempPath), Row(3, 30) :: Row(4, 40) :: Nil)
   }
@@ -62,14 +62,14 @@ class DeleteScalaSuite extends DeleteSuiteBase {
       }
     }
 
-    val deltaTable: io.delta.DeltaTable = {
+    val deltaTable: DeltaTable = {
       val (tableNameOrPath, optionalAlias) = parse(target)
       val isPath: Boolean = tableNameOrPath.startsWith("delta.")
       val table = if (isPath) {
         val path = tableNameOrPath.stripPrefix("delta.`").stripSuffix("`")
-        io.delta.DeltaTable.forPath(spark, path)
+        io.delta.tables.DeltaTable.forPath(spark, path)
       } else {
-        io.delta.DeltaTable(spark.table(tableNameOrPath))
+        io.delta.tables.DeltaTable(spark.table(tableNameOrPath))
       }
       optionalAlias.map(table.as(_)).getOrElse(table)
     }
